test: annotation_specs
